+++ tests/python/openassetio/hostApi/test_manager.py
@@ -219,16 +219,6 @@ class Test_Manager_initialize:
         mock_manager_interface.mock.initialize.assert_called_once_with(a_dict, a_host_session)
 
 
-class Test_Manager_prefetch:
-
-    def test_wraps_the_corresponding_method_of_the_held_interface(
-            self, manager, mock_manager_interface, a_host_session, some_refs, a_context):
-
-        method = mock_manager_interface.mock.prefetch
-        assert manager.prefetch(some_refs, a_context) == method.return_value
-        method.assert_called_once_with(some_refs, a_context, a_host_session)
-
-
 class Test_Manager_flushCaches:
 
     def test_wraps_the_corresponding_method_of_the_held_interface(